+++ DiscreteLib/EquationId/EquationIdStorage.h
@@ -19,8 +19,8 @@ public:
     virtual void set(size_t pt_id, long eqs_id) = 0;
     virtual size_t setAll(size_t address_start, size_t dn_pt=1) = 0;
     virtual size_t size() const = 0;
-    virtual long address(size_t key_id) const = 0;
-    virtual long key(long address_id) const = 0;
+    virtual size_t address(size_t key_id) const = 0;
+    virtual size_t key(size_t address_id) const = 0;
     virtual bool isSequential() const = 0;
 };
